+++ incs/ray.h
@@ -12,7 +12,7 @@ typedef struct s_window	t_window;
 
 typedef struct s_hit_data 
 {
-	float	t;
+	double	t;
 	int		is_front;
 	t_p3	hit;
 	t_vec3	normal;
@@ -27,7 +27,7 @@ typedef struct s_ray
 	t_vec3		direction;
 	t_vec3		ray;
 	t_color		rgb;
-	float		length;
+	double		length;
 	t_vec3		norm;
 	t_interval	lim;
 }	t_ray;
@@ -38,6 +38,6 @@ t_ray  			init_ray(t_vec3 *origin, t_vec3 *dir);
 t_color			tst_ray_color(const t_ray *r);
 t_color			ray_color(const t_ray *r, int max_depth);
 t_ray			get_ray(t_window *win, t_camera *camera, t_ivec2 *pix_pos);
-t_vec3			at(t_ray *r, float t);
+t_vec3			at(t_ray *r, double t);
 
 #endif
