+++ srcs/colours-utils.c
@@ -6,7 +6,7 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 14:45:13 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/18 18:06:16 by hmontoya         ###   ########.fr       */
+/*   Updated: 2025/01/16 18:24:04 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,9 +45,9 @@ t_color	vec3_to_rgb(t_vec3 v3)
  * Maps the a dimention into another by scaling it.
  *  
 */
-int lerpRGB(float pos, t_color from, t_color to)
+int lerpRGB(double pos, t_color from, t_color to)
 {
-	float		a;
+	double		a;
 	t_ncolor	f;
 	t_ncolor	t;
 	t_color 	i;
