+++ srcs/math-interval-absolutes.c
@@ -6,7 +6,7 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 14:47:34 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/17 11:39:45 by hmontoya         ###   ########.fr       */
+/*   Updated: 2025/01/16 18:24:04 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void empty(t_interval *inter)
 	inter->max = -INFINITY;	
 }
 
-float clamp(t_interval *inter, float x) 
+double clamp(t_interval *inter, double x) 
 {
 	if (x < inter->min)
 		return inter->min;
