+++ srcs/vec3-float-operations.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   vec3-float-operations.c                           :+:      :+:    :+:   */
+/*   vec3-float-operations.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hmontoya <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 14:51:38 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/16 14:51:47 by hmontoya         ###   ########.fr       */
+/*   Updated: 2025/01/16 18:24:04 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vectors.h"
 
-t_vec3	scale_v3(t_vec3 v1, float t)
+t_vec3	scale_v3(t_vec3 v1, double t)
 {
 	t_vec3	new;
 
@@ -22,7 +22,7 @@ t_vec3	scale_v3(t_vec3 v1, float t)
 	return (new);
 }
 
-t_vec3	div_v3_dbl(t_vec3 v1, float t)
+t_vec3	div_v3_dbl(t_vec3 v1, double t)
 {
 	t_vec3	new;
 
@@ -32,7 +32,7 @@ t_vec3	div_v3_dbl(t_vec3 v1, float t)
 	return (new);
 }
 
-t_vec3	sum_v3_dbl(t_vec3 v1, float t)
+t_vec3	sum_v3_dbl(t_vec3 v1, double t)
 {
 	t_vec3	new;
 
@@ -42,7 +42,7 @@ t_vec3	sum_v3_dbl(t_vec3 v1, float t)
 	return (new);
 }
 
-t_vec3	sub_v3_dbl(t_vec3 v1, float t)
+t_vec3	sub_v3_dbl(t_vec3 v1, double t)
 {
 	t_vec3 new;
 
@@ -52,8 +52,8 @@ t_vec3	sub_v3_dbl(t_vec3 v1, float t)
 	return (new);
 }
 
-float	vec3_to_dbl(t_vec3 v)
+double	vec3_to_dbl(t_vec3 v)
 {
 	return (v.x + v.y + v.z);
 }
-//return ((float)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
+//return ((double)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
