+++ srcs/vec3-operations.c
@@ -6,13 +6,13 @@
 /*   By: hmontoya <hmontoya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 14:52:12 by hmontoya          #+#    #+#             */
-/*   Updated: 2024/12/30 16:56:44 by hmontoya         ###   ########.fr       */
+/*   Updated: 2025/01/16 18:24:04 by hmontoya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vectors.h"
 
-t_vec3	vec3(float x, float y, float z)
+t_vec3	vec3(double x, double y, double z)
 {
 	t_vec3	v3;
 
@@ -41,4 +41,4 @@ t_vec3	sub_v3(t_vec3 v1, t_vec3 v2)
 	new.z = v1.z - v2.z;
 	return (new);
 }
-//return ((float)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
+//return ((double)(v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
